# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RVC2API Backend",
    "description": "Modernized backend API for RV-C CANbus monitoring and control",
    "version": "2.0.0"
  },
  "paths": {
    "/api/entities": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "List entities",
        "description": "Return all entities, optionally filtered by device_type and/or area.",
        "operationId": "list_entities_api_entities_get",
        "parameters": [
          {
            "name": "device_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity device_type",
              "title": "Device Type"
            },
            "description": "Filter by entity device_type"
          },
          {
            "name": "area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity suggested_area",
              "title": "Area"
            },
            "description": "Filter by entity suggested_area"
          }
        ],
        "responses": {
          "200": {
            "description": "Dictionary of entities matching the filter criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Entities Api Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entities/{entity_id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get entity details",
        "description": "Return the latest value for one entity.",
        "operationId": "get_entity_api_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The entity object with current state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Entity Api Entities  Entity Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/entities/{entity_id}/history": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get entity history",
        "description": "Return the history of state changes for a specific entity.",
        "operationId": "get_entity_history_api_entities__entity_id__history_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of history entries to return",
              "title": "Limit"
            },
            "description": "Maximum number of history entries to return"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Unix timestamp to filter history from",
              "title": "Since"
            },
            "description": "Unix timestamp to filter history from"
          }
        ],
        "responses": {
          "200": {
            "description": "List of historical state entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Entity History Api Entities  Entity Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/entities/{entity_id}/control": {
      "post": {
        "tags": [
          "entities"
        ],
        "summary": "Control entity",
        "description": "Send a control command to an entity.",
        "operationId": "control_entity_api_entities__entity_id__control_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Control command response with execution details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlEntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ControlCommand": {
        "properties": {
          "command": {
            "type": "string",
            "title": "Command"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State",
            "description": "Target state: 'on' or 'off'. Required only for 'set' command."
          },
          "brightness": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Brightness",
            "description": "Brightness percent (0-100). Only used when command is 'set' and state is 'on'."
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "ControlCommand",
        "description": "Defines the structure for sending control commands to an entity, typically a light."
      },
      "ControlEntityResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "command": {
            "type": "string",
            "title": "Command"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "brightness": {
            "type": "integer",
            "title": "Brightness"
          },
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "status",
          "entity_id",
          "command",
          "state",
          "brightness",
          "action"
        ],
        "title": "ControlEntityResponse",
        "description": "Response model for individual entity control commands, confirming the action taken."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}